Avoid parsing of hostname from router string by setting _client on CLI.
* Importing an image file back into omero via import in omero.cli.CLI
requires setting a session UUID and specifying the hostname (not always
'localhost' since the OMERO.grid can be quite complex).
* It should be possible to simply pass a copy of the client to the CLI
object, and avoid such configuration. Unfortunately, there is a bug that
prevented the 'omero.host' property to be configured.
* This commit works around this bug by setting the property ourselves
(fix from Josh Moore -- see link to the mailing list on source).
* the previous solution created a process with the session UUID and
hostname on the name which made it very easy for anyone listening
to take control of the session. This solution also fixes it.