Fix clean-up of temporary files and finish matlab_block.
* fix cleanup of temporary files by actually close the file descriptor and
then double check that the file was actually removed.  Empyting the list
was not enough since there were copies of the objects on fin and fout (we
may want to use fin and fout to store the names only and go back to simply
empty the list which would allow to keep files between blocks by making a
copy of the file descriptor).
* use single '.tiff' instead of '.ome.tiff' when creating file because it
confuses code that split filename based on the last dot.
* fix syntax errors on the rest of matlab_block class to make it work.